# Improve perf of Graph.Scope.scopeFor hotspot