--- a/engine/runtime-compiler/src/main/scala/org/enso/compiler/context/LocalScope.scala
+++ b/engine/runtime-compiler/src/main/scala/org/enso/compiler/context/LocalScope.scala
@@ -140,10 +140,10 @@ class LocalScope(
       .getOrElse(Map())
 
     scope.occurrences.foreach {
-      case x: AliasGraph.Occurrence.Def =>
+      case (id, x: AliasGraph.Occurrence.Def) =>
         parentResult += x.symbol -> new FramePointer(
           level,
-          allFrameSlotIdxs(x.id)
+          allFrameSlotIdxs(id)
         )
       case _ =>
     }